chore: Breaks up ThemeProvider, removes empty css, cleans up imports
