+++ packages/components/lib/contexts/index.ts
@@ -1,0 +1,1 @@
+export * from './theme';
