+++ packages/components/lib/contexts/theme/ThemeContext.tsx
@@ -1,0 +1,12 @@
+import { createContext, useContext } from 'react';
+import { Theme, ThemeContextType } from './types';
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
+
+export const useTheme = (): ThemeContextType => {
+  const context = useContext(ThemeContext);
+  if (!context) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+  return context;
+};
