+++ packages/components/lib/contexts/theme/index.ts
@@ -1,0 +1,3 @@
+export { useTheme } from './ThemeContext';
+export { default as LazyOllamaWebThemeProvider } from './ThemeProvider';
+export { type Theme, type ThemeContextType } from './types';
