+++ packages/components/lib/contexts/theme/types.ts
@@ -1,0 +1,8 @@
+type Theme = 'light' | 'dark';
+
+interface ThemeContextType {
+  theme: Theme;
+  toggleTheme: () => void;
+}
+
+export { Theme, ThemeContextType };
