--- packages/components/lib/themes/ThemeProvider.tsx
@@ -1,60 +1,0 @@
-import React, { createContext, useState, useContext, useEffect, ReactNode } from 'react';
-import '../styles/variables.css';
-
-type Theme = 'light' | 'dark';
-
-interface ThemeContextType {
-  theme: Theme;
-  toggleTheme: () => void;
-}
-
-const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
-
-export const useTheme = (): ThemeContextType => {
-  const context = useContext(ThemeContext);
-  if (!context) {
-    throw new Error('useTheme must be used within a ThemeProvider');
-  }
-  return context;
-};
-
-interface ThemeProviderProps {
-  children: ReactNode;
-  initialTheme?: Theme;
-}
-
-export const ThemeProvider: React.FC<ThemeProviderProps> = ({
-  children,
-  initialTheme = 'light'
-}) => {
-  const [theme, setTheme] = useState<Theme>(() => {
-    // Check for saved theme preference in localStorage
-    const savedTheme = localStorage.getItem('theme') as Theme | null;
-
-    // Check for system preference
-    if (!savedTheme && window.matchMedia) {
-      const prefersDark = window.matchMedia('(prefers-color-scheme: dark)').matches;
-      return prefersDark ? 'dark' : 'light';
-    }
-
-    return savedTheme || initialTheme;
-  });
-
-  useEffect(() => {
-    // Update data-theme attribute on document
-    document.documentElement.setAttribute('data-theme', theme);
-
-    // Save theme preference to localStorage
-    localStorage.setItem('theme', theme);
-  }, [theme]);
-
-  const toggleTheme = () => {
-    setTheme((prevTheme) => (prevTheme === 'light' ? 'dark' : 'light'));
-  };
-
-  return (
-    <ThemeContext.Provider value={{ theme, toggleTheme }}>{children}</ThemeContext.Provider>
-  );
-};
-
-export default ThemeProvider;
