+++ packages/components/package.json
@@ -45,6 +45,7 @@
     "typescript": "^5.0.0"
   },
   "dependencies": {
+    "classnames": "^2.5.1",
     "react-syntax-highlighter": "^15.6.1",
     "sleepydogs": "1.0.8-alpha.2"
   }
